requirements.txt
coloredlogs==10.0
humanfriendly==4.18
paho-mqtt==1.4.0
python-dotenv==0.10.3
